+++ src/pages/dao/settings/SafeSettingsPage.tsx
@@ -1,7 +1,7 @@
 import { Flex, useBreakpointValue } from '@chakra-ui/react';
 import { useTranslation } from 'react-i18next';
 import { Outlet, useLocation } from 'react-router-dom';
-import SettingsNavigation from '../../../components/SafeSettings/SettingsNavigation';
+import { SettingsNavigation } from '../../../components/SafeSettings/SettingsNavigation';
 import PageHeader from '../../../components/ui/page/Header/PageHeader';
 import { useCurrentDAOKey } from '../../../hooks/DAO/useCurrentDAOKey';
 import { useStore } from '../../../providers/App/AppProvider';
